+++ inst/include/dplyr/Result/CallProxy.h
@@ -37,6 +37,7 @@ namespace dplyr {
                 for( int i=0; i<n; i++){
                     proxies[i].set( subsets[proxies[i].symbol] ) ;     
                 }
+            
                 Shield<SEXP> res( call.eval(env) ) ;
                 return res ;
             } else if( TYPEOF(call) == SYMSXP) {